refactor: register harness compilers by type
